move new e2e into their own module
